requirements.txt
aiohttp==3.5.4
aiomysql==0.0.20
amqp==2.5.0
apply==1.5
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.8.0
bitstring==3.1.6
bytebuffer==0.1.2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
coverage==4.5.4
cryptography==2.7
dnspython==1.16.0
eq==0.0.10
etcd3==0.10.0
eventlet==0.25.0
files==1.1.1
Flask==1.1.1
gevent==1.4.0
google==2.0.2
greenlet==0.4.15
grequests==0.4.0
grpcio==1.23.0
gsocketpool==0.1.6
idna==2.8
idna-ssl==1.1.0
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.3
kombu==4.6.3
MarkupSafe==1.1.1
mock==3.0.5
monotonic==1.5
multidict==4.5.2
nameko==2.12.0
path.py==12.0.1
protobuf==3.9.1
py-etcd==0.1.4
pycparser==2.19
pymongo==3.9.0
PyMySQL==0.9.2
python-consul==1.1.0
PyYAML==5.1.1
pyyml==0.0.2
redis==3.0.0
redis-py-cluster==2.0.0
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
SQLAlchemy==1.3.6
tenacity==5.1.1
typing-extensions==3.7.4
urllib3==1.25.3
vine==1.3.0
Werkzeug==0.15.5
wrapt==1.11.2
yarl==1.3.0
yml==0.0.1
zipp==0.5.2
